+++ probe-rs/src/bin/probe-rs/util/rtt.rs
@@ -25,6 +25,7 @@ pub fn attach_to_rtt(
     elf_file: &Path,
     rtt_config: &RttConfig,
     timestamp_offset: UtcOffset,
+    log_format: Option<&str>,
 ) -> Result<RttActiveTarget, anyhow::Error> {
     log::info!("Initializing RTT");
     let rtt_header_address = if let Ok(mut file) = File::open(elf_file) {
@@ -47,7 +48,8 @@ pub fn attach_to_rtt(
     match Rtt::attach_region(core, memory_map, &rtt_header_address) {
         Ok(rtt) => {
             log::info!("RTT initialized.");
-            let app = RttActiveTarget::new(rtt, elf_file, rtt_config, timestamp_offset)?;
+            let app =
+                RttActiveTarget::new(rtt, elf_file, rtt_config, timestamp_offset, log_format)?;
             Ok(app)
         }
         Err(err) => Err(anyhow!("Error attempting to attach to RTT: {}", err)),
@@ -243,7 +245,7 @@ impl RttActiveChannel {
     pub fn get_rtt_data(
         &mut self,
         core: &mut Core,
-        defmt_state: Option<&(defmt_decoder::Table, Option<defmt_decoder::Locations>)>,
+        defmt_state: Option<&DefmtState>,
     ) -> Result<Option<(String, String)>, anyhow::Error> {
         self
             .poll_rtt(core)
@@ -271,33 +273,21 @@ impl RttActiveChannel {
                             }
                             DataFormat::Defmt => {
                                 match defmt_state {
-                                    Some((table, locs)) => {
+                                    Some(DefmtState { table, locs, formatter }) => {
                                         let mut stream_decoder = table.new_stream_decoder();
                                         stream_decoder.received(&self.rtt_buffer.0[..bytes_read]);
                                         loop {
                                             match stream_decoder.decode() {
                                                 Ok(frame) => {
                                                     let loc = locs.as_ref().and_then(|locs| locs.get(&frame.index()) );
-                                                    writeln!(formatted_data, "{}", frame.display(true)).map_or_else(|err| log::error!("Failed to format RTT data - {:?}", err), |r|r);
-                                                    if self.show_location {
-                                                        if let Some(loc) = loc {
-                                                            let relpath = if let Ok(relpath) =
-                                                                loc.file.strip_prefix(&std::env::current_dir().unwrap())
-                                                            {
-                                                                relpath
-                                                            } else {
-                                                                // not relative; use full path
-                                                                &loc.file
-                                                            };
-                                                            writeln!(formatted_data,
-                                                                "└─ {}:{}",
-                                                                relpath.display(),
-                                                                loc.line
-                                                            ).map_or_else(|err| log::error!("Failed to format RTT data - {:?}", err), |r|r);
-                                                        } else {
-                                                            writeln!(formatted_data, "└─ <invalid location: defmt frame-index: {}>", frame.index()).map_or_else(|err| log::error!("Failed to format RTT data - {:?}", err), |r|r);
-                                                        }
-                                                    }
+                                                    let (file, line, module) = if let Some(loc) = loc {
+                                                        let relpath = loc.file.strip_prefix(&std::env::current_dir().unwrap()).unwrap_or(&loc.file);
+                                                        (Some(relpath.display().to_string()), Some(loc.line.try_into().unwrap()), Some(loc.module.as_str()))
+                                                    } else {
+                                                        (Some(format!("└─ <invalid location: defmt frame-index: {}>", frame.index())), None, None)
+                                                    };
+                                                    let s = formatter.format_to_string(frame, file.as_deref(), line, module);
+                                                    writeln!(formatted_data, "{s}")?;
                                                     continue;
                                                 },
                                                 Err(DecodeError::UnexpectedEof) => break,
@@ -341,7 +331,14 @@ impl RttActiveChannel {
 #[derive(Debug)]
 pub struct RttActiveTarget {
     pub active_channels: Vec<RttActiveChannel>,
-    pub defmt_state: Option<(defmt_decoder::Table, Option<defmt_decoder::Locations>)>,
+    pub defmt_state: Option<DefmtState>,
+}
+
+#[derive(Debug)]
+pub struct DefmtState {
+    table: defmt_decoder::Table,
+    locs: Option<defmt_decoder::Locations>,
+    formatter: defmt_decoder::log::Formatter,
 }
 
 impl RttActiveTarget {
@@ -351,6 +348,7 @@ impl RttActiveTarget {
         elf_file: &Path,
         rtt_config: &RttConfig,
         timestamp_offset: UtcOffset,
+        log_format: Option<&str>,
     ) -> Result<Self> {
         let mut active_channels = Vec::new();
         // For each channel configured in the RTT Control Block (`Rtt`), check if there are additional user configuration in a `RttChannelConfig`. If not, apply defaults.
@@ -403,7 +401,29 @@ impl RttActiveTarget {
                     err
                 )
             })?;
+
+            let show_location = active_channels
+                .get(0)
+                .expect("`active_channels` is not empty")
+                .show_location;
+
             if let Some(table) = defmt_decoder::Table::parse(&elf)? {
+                let has_timestamp = table.has_timestamp();
+
+                // Format options:
+                // 1. Custom format
+                // 2. Default with timestamp with location
+                // 3. Default with timestamp without location
+                // 4. Default without timestamp with location
+                // 5. Default without timestamp without location
+                let format = log_format.unwrap_or(match (show_location, has_timestamp) {
+                    (true, true) => "{t} {L} {s}\n└─ {m} @ {F}:{l}",
+                    (true, false) => "{L} {s}\n└─ {m} @ {F}:{l}",
+                    (false, true) => "{t} {L} {s}",
+                    (false, false) => "{L} {s}",
+                });
+                let formatter = defmt_decoder::log::Formatter::new(format);
+
                 let locs = {
                     let locs = table.get_locations(&elf)?;
 
@@ -419,7 +439,11 @@ impl RttActiveTarget {
                         None
                     }
                 };
-                Some((table, locs))
+                Some(DefmtState {
+                    table,
+                    locs,
+                    formatter,
+                })
             } else {
                 log::warn!("No `Table` definition in DWARF info; compile your program with `debug = 2` to enable location info.");
                 None